+++ lib/src/analyze.dart
@@ -1,5 +1,3 @@
-import 'dart:io';
-
 import 'package:dart_pre_commit/src/file_resolver.dart';
 import 'package:meta/meta.dart';
 import 'package:path/path.dart';