+++ test/file_resolver_test.dart
@@ -6,6 +6,7 @@ import 'package:test/test.dart';
 
 void main() {
   Directory testDir;
+  final sut = FileResolver();
 
   setUp(() async {
     testDir = await Directory.systemTemp.createTemp();
@@ -14,6 +15,13 @@ void main() {
   tearDown(() async {
     await testDir.delete(recursive: true);
     testDir = null;
+  });
+
+  test("exists checks if file exists", () async {
+    await File(join(testDir.path, "test1.dart")).create();
+
+    expect(await sut.exists(join(testDir.path, "test1.dart")), true);
+    expect(await sut.exists(join(testDir.path, "test2.dart")), false);
   });
 
   test("resolve resolves symlinks to a relative path", () async {
@@ -29,7 +37,6 @@ void main() {
         recursive: true,
       );
 
-      final sut = FileResolver();
       final resPath = await sut.resolve(l1.absolute.path);
       expect(resPath, join("dir1", "file1.dart"));
     } finally {
@@ -60,7 +67,6 @@ void main() {
         recursive: true,
       );
 
-      final sut = FileResolver();
       final resPath = await sut.resolveAll([
         l1.absolute.path,
         l2.absolute.path,