+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - `dart analyze` now treats lint infos as errors
 - Replace legacy dart tools in README
+- Add missing await in program runner
 
 ## [1.0.0] - 2020-10-07
 ### Added